01CloudFunctions/a2_getentidades/requirements.txt
# Function dependencies, for example:
google-cloud-storage
pandas >= 1.1.0
python-dateutil==2.8.1
beautifulsoup4==4.10.0
numpy==1.21.1
DateTime==4.3
requests